% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recup_palette.R
\name{recup_palette}
\alias{recup_palette}
\title{Returns a palette of the graphic chart of INSEE}
\usage{
recup_palette(stylePalette, nbNeg = 0, nbPos = 0)
}
\arguments{
\item{stylePalette}{chaine de caracteres (character). A choisir parmi
"defaut", "Insee_Rouge", "Insee_Jaune", "Insee_Bleu", "Insee_Turquoise",
"Insee_Vert", "Insee_Violet" ou "Insee_Gris".}

\item{nbNeg}{numeric. Nombre de classes negatives. Par defaut 0.}

\item{nbPos}{numeric. Nombre de classes positives. Par defaut 0.}
}
\value{
Retourne une liste d'un vecteur caracteres.
}
\description{
Returns a palette of the graphic chart of INSEE.
}
\details{
Les palettes de couleurs disponibles sont celles de la Charte Graphique
INSEE (hors palette "defaut"). En exécutant la fonction affiche_palette(nomPalette),
il est possible de visualiser les couleurs de chaque palette disponible.
Les modalités pour l'argument 'nomPalette' sont "defaut", "Insee_Rouge", "Insee_Jaune", 
"Insee_Bleu", "Insee_Turquoise", "Insee_Vert", "Insee_Violet" et "Insee_Gris".

Seules les palettes "Insee_Rouge" et "Insee_Jaune" proposent des palettes pour des valeurs
négatives et positives. Pour ces deux palettes, la palette "Insee_Bleu" inversée est
utilisée pour représenter les valeurs négatives.

La fonction renvoie une liste d'un vecteur composé du nom de la palette au format
nomPalette_xN_yP et des valeurs hexadécimales de la palette.

La liste des valeurs hexadécimales de toutes les palettes est disponible en
chargeant les données suivantes : data("palettes_insee").

Il est obligatoire de spécifier un nombre de classes négatives ou positives
supérieur à 0.
}
\examples{

recup_palette(stylePalette = "defaut", nbNeg = 3, nbPos = 3)

# $defaut_3N3P
# [1] "#003269" "#005289" "#95BAE2" "#E4A75A" "#D47130" "#B24B1D"

}
\references{
Un convertisseur de couleurs pour visualiser une couleur a
partir de son nom, son code hexadecimal ou RGB :
http://www.proftnj.com/RGB3.htm
}
\seealso{
\code{\link{set_couleur_classes}}
}
\keyword{documentation}
