% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiche_palette.R
\name{affiche_palette}
\alias{affiche_palette}
\title{Returns a palette colors of the graphic chart of INSEE}
\usage{
affiche_palette(nomPalette = "Bleu_Jaune")
}
\arguments{
\item{nomPalette}{chaine de caracteres (character). A choisir parmi
"Bleu_Jaune", "Bleu_Rouge", "Bleu", "Jaune", "Gris", "Turquoise", "Vert"
ou "Violet".}
}
\value{
Affiche la palette dans l'onglet Plots.
}
\description{
Returns a palette colors of the graphic chart of INSEE.
}
\details{
Les palettes de couleurs disponibles sont celles de la Charte Graphique
INSEE. En exécutant la fonction affiche_palette(nomPalette), il est possible de
visualiser les couleurs de chaque palette disponible. Les modalités pour
l'argument 'nomPalette' sont "Bleu_Jaune", "Bleu_Rouge", "Bleu", "Jaune",
"Gris", "Turquoise", "Vert" et "Violet".

La fonction permet de visualiser les couleurs de chaque palette.

La liste des valeurs hexadécimales de toutes les palettes est dipsonible en
chargeant les données suivantes : data("palettes_insee")
}
\examples{

affiche_palette(nomPalette = "Bleu_Jaune")

}
\references{
Un convertisseur de couleurs pour visualiser une couleur a
partir de son nom, son code hexadecimal ou RGB :
http://www.proftnj.com/RGB3.htm
}
\seealso{
\code{\link{set_couleur_classes}}
}
\keyword{documentation}
