% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanis-package.R
\docType{data}
\name{donnees}
\alias{donnees}
\alias{com_dep_13_30_83_84}
\alias{donnees_a_facon}
\alias{donnees_monoloc}
\alias{donnees_biloc}
\alias{donnees_biloc_saphirs}
\alias{depm}
\alias{regm}
\alias{fram}
\title{Examples of data}
\format{\code{com_dep_13_30_83_84} Un data frame de 792 observations et 3
variables.  \describe{ \item{list("CODE")}{vecteur caracteres}
\item{list("LIBELLE")}{vecteur caracteres}
\item{list("geometry")}{sfc_GEOMETRY} }

Un data frame de 281 observations et 3 variables.  \describe{
\item{list("DEPCOM")}{vecteur caracteres} \item{list("ZE2010")}{vecteur
caracteres} \item{list("LIB_ZE2010")}{vecteur caracteres} }

\code{donnees_monoloc} Un data frame de 96 observations et 6 variables.
\describe{ \item{list("COD_DEP")}{vecteur caracteres}
\item{list("LIB_DEP")}{vecteur caracteres} \item{list("POP_2010")}{vecteur
numerique} \item{list("POP_2015")}{vecteur numerique}
\item{list("VAR_AN_MOY")}{vecteur numerique} \item{list("REG")}{vecteur
caracteres} }

\code{donnees_biloc} Un data frame de 378 observations et 3 variables.
\describe{ \item{list("REG_DEPART")}{vecteur caracteres}
\item{list("REG_ARRIVEE")}{vecteur caracteres} \item{list("MIGR")}{vecteur
numerique} }

\code{donnees_biloc_saphirs} Un data frame de 21 observations et 3
variables.  \describe{ \item{list("REG_DEPART")}{vecteur caracteres}
\item{list("REG_ARRIVEE")}{vecteur caracteres} \item{list("MIGR")}{vecteur
numerique} }

\code{depm} Un data frame de 96 observations et 5 variables.  \describe{
\item{list("CODE")}{vecteur caracteres} \item{list("LIBELLE")}{vecteur
caracteres} \item{list("REG")}{vecteur caracteres}
\item{list("SURF")}{vecteur numerique} \item{list("geometry")}{sfc_GEOMETRY}
}

\code{regm} Un data frame de 13 observations et 4 variables.  \describe{
\item{list("CODE")}{vecteur caracteres} \item{list("LIBELLE")}{vecteur
caracteres} \item{list("SURF")}{vecteur numerique}
\item{list("geometry")}{sfc_GEOMETRY} }

\code{fram} Un data frame de 1 observation et 4 variables.  \describe{
\item{list("CODE")}{vecteur caracteres} \item{list("LIBELLE")}{vecteur
caracteres} \item{list("SURF")}{vecteur numerique}
\item{list("geometry")}{sfc_GEOMETRY} }}
\usage{
data("com_dep_13_30_83_84")

data("donnees_a_facon")

data("donnees_monoloc")

data("donnees_biloc")

data("donnees_biloc_saphirs")

data("depm")

data("regm")

data("fram")
}
\description{
Data tables for examples.

- \code{com_dep_13_30_83_84} for \code{zonage_a_facon}

- \code{donnees_a_facon} for \code{zonage_a_facon}

- \code{donnees_monoloc} for proportional circles, chroropleth and typology

- \code{donnees_biloc} for propotionnal arrows

- \code{donnees_biloc_saphirs} for sapphire arrows

- \code{depm, regm, fram} for the meshes and contours of the maps
}
\examples{

data(com_dep_13_30_83_84)
data(donnees_a_facon)
data(donnees_monoloc)
data(donnees_biloc)
data(donnees_biloc_saphirs)
data(depm)
data(regm)
data(fram)

}
\keyword{datasets}
