% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanis-package.R
\docType{data}
\name{donnees}
\alias{donnees}
\alias{com_dep_13_30_83_84}
\alias{donnees_a_facon}
\alias{donnees_monoloc}
\alias{donnees_biloc}
\alias{donnees_biloc_saphirs}
\alias{depm}
\alias{regm}
\alias{fram}
\alias{palettes_insee}
\title{Examples of data}
\format{
\code{com_dep_13_30_83_84} Un data frame de 792 observations et 3
variables.  \describe{ \item{"CODE"}{vecteur caracteres}
\item{"LIBELLE"}{vecteur caracteres}
\item{"geometry"}{sfc_GEOMETRY} }

Un data frame de 281 observations et 3 variables.  \describe{
\item{"DEPCOM"}{vecteur caracteres} \item{"ZE2010"}{vecteur
caracteres} \item{"LIB_ZE2010"}{vecteur caracteres} }

\code{donnees_monoloc} Un data frame de 96 observations et 6 variables.
\describe{ \item{"COD_DEP"}{vecteur caracteres}
\item{"LIB_DEP"}{vecteur caracteres} \item{"POP_2010"}{vecteur
numerique} \item{"POP_2015"}{vecteur numerique}
\item{"VAR_AN_MOY"}{vecteur numerique} \item{"REG"}{vecteur
caracteres} }

\code{donnees_biloc} Un data frame de 378 observations et 3 variables.
\describe{ \item{"REG_DEPART"}{vecteur caracteres}
\item{"REG_ARRIVEE"}{vecteur caracteres} \item{"MIGR"}{vecteur
numerique} }

\code{donnees_biloc_saphirs} Un data frame de 21 observations et 3
variables.  \describe{ \item{"REG_DEPART"}{vecteur caracteres}
\item{"REG_ARRIVEE"}{vecteur caracteres} \item{"MIGR"}{vecteur
numerique} }

\code{depm} Un data frame de 96 observations et 5 variables.  \describe{
\item{"CODE"}{vecteur caracteres} \item{"LIBELLE"}{vecteur
caracteres} \item{"REG"}{vecteur caracteres}
\item{"SURF"}{vecteur numerique} \item{"geometry"}{sfc_GEOMETRY}
}

\code{regm} Un data frame de 13 observations et 4 variables.  \describe{
\item{"CODE"}{vecteur caracteres} \item{"LIBELLE"}{vecteur
caracteres} \item{"SURF"}{vecteur numerique}
\item{"geometry"}{sfc_GEOMETRY} }

\code{fram} Un data frame de 1 observation et 4 variables.  \describe{
\item{"CODE"}{vecteur caracteres} \item{"LIBELLE"}{vecteur
caracteres} \item{"SURF"}{vecteur numerique}
\item{"geometry"}{sfc_GEOMETRY} }

\code{palettes_insee} Une liste de vecteurs caracteres.  \describe{
\item{"nomPalette"}{vecteur caracteres de codes hexadecimaux} }
}
\usage{
data("com_dep_13_30_83_84")

data("donnees_a_facon")

data("donnees_monoloc")

data("donnees_biloc")

data("donnees_biloc_saphirs")

data("depm")

data("regm")

data("fram")

data("palettes_insee")
}
\description{
Data tables for examples.

- \code{com_dep_13_30_83_84} for \code{zonage_a_facon}

- \code{donnees_a_facon} for \code{zonage_a_facon}

- \code{donnees_monoloc} for proportional circles, chroropleth and typology

- \code{donnees_biloc} for propotionnal arrows

- \code{donnees_biloc_saphirs} for sapphire arrows

- \code{depm, regm, fram} for the meshes and contours of the maps

- \code{palettes_insee} for palette of the graphic chart of INSEE
}
\examples{

data(com_dep_13_30_83_84)
data(donnees_a_facon)
data(donnees_monoloc)
data(donnees_biloc)
data(donnees_biloc_saphirs)
data(depm)
data(regm)
data(fram)
data(palettes_insee)

}
\keyword{datasets}
