% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recup_palette.R
\name{recup_palette}
\alias{recup_palette}
\title{Returns a palette of the graphic chart of INSEE}
\usage{
recup_palette(stylePalette, nbNeg = 0, nbPos = 0)
}
\arguments{
\item{stylePalette}{chaine de caracteres (character). A choisir parmi
"Bleu_Jaune", "Bleu_Rouge", "Bleu", "Jaune", "Gris", "Turquoise", "Vert"
ou "Violet".}

\item{nbNeg}{numeric. Nombre de classes negatives. Par defaut 0.}

\item{nbPos}{numeric. Nombre de classes positives. Par defaut 0.}
}
\value{
Retourne une liste d'un vecteur caracteres.
}
\description{
Returns a palette of the graphic chart of INSEE.
}
\details{
Les palettes de couleurs disponibles sont celles de la Charte Graphique
INSEE. En exécutant la fonction affiche_palette(nomPalette), il est possible de
visualiser les couleurs de chaque palette disponible. Les modalités pour
l'argument 'nomPalette' sont "Bleu_Jaune", "Bleu_Rouge", "Bleu", "Jaune",
"Gris", "Turquoise", "Vert" et "Violet".

La fonction renvoie une liste d'un vecteur composé du nom de la palette au format
nomPalette_xN_yP et des valeurs hexadécimales de la palette.

La liste des valeurs hexadécimales de toutes les palettes est disponible en
chargeant les données suivantes : data("palettes_insee").

Il est obligatoire de spécifier un nombre de classes négatives ou positives
supérieur à 0.
}
\examples{

recup_palette(stylePalette = "Bleu_Jaune", nbNeg = 3, nbPos = 3)

# $Bleu_Jaune_3N3P
# [1] "#0F417A" "#286AC7" "#8DB0E1" "#FBD616" "#FFC300" "#E89E0B"

}
\references{
Un convertisseur de couleurs pour visualiser une couleur a
partir de son nom, son code hexadecimal ou RGB :
http://www.proftnj.com/RGB3.htm
}
\seealso{
\code{\link{set_couleur_classes}}
}
\keyword{documentation}
